Maps instead of arrays; pointers for values
For faster lookup and cleaner code, change lists of commonly accessed items
to be maps. Turns out a map[int]bool is the same size as []int, which is a
bonus for scaling. Things like lists of dendrites and axons, for example.

There was a serious problem during activation of the cells in the network.
Turns our we need to use pointers in maps instead of the actual values.
For some reason you cannot call a method on a struct in a map; only on a
pointer to it. This was a hard to find bug because it was setting synapse
dendrite and axon IDs to 0, a technically valid value. I guess copying
them cleared the value. If it had not, we may have had serious memory
leakage and the network would not really get trained, but everything
would appear to work. So..good job, Go team.