Fix issue which may have been causing concurrent map read/write crash.
Wrapping the cell AP fire synapse loop in a goroutine seems to help fix
an issue where the program crashed due to simultaneous read/write ops
on a synapse map. Most likely, the network was disabled but still firing
and we started growing it and operating heavily on the synapse map.

We also re-check that the network is not disabled right before firing a
synapse to make the firings fizzle out faster and save CPU.