wip: large adjustment wit charrnn enables generic potential.Train()
Training is somewhat generic now, with an implementation like a classifier
or character-level NN  needing to provide their own data setup. The
potential lib can run the trainings and sampling while not caring about
what kind of data is being used. It is all just inputs and outputs now.

However:
- shake.go is surely broken
- unsure if sampling works (needs tests!)
- unsure if training will be able to get ported all the way (unported and untested)