Fix two bugs related to applying the diff.
1. When copying a synapse to the network, and the synapse ID already existed,
and one or more of its cells were new, we need to delete its old synapse ID
from its new cells. That was causing an integrity issue very infrequently,
only visible after a Prune eventually cleared out the oldly referenced synapse,
which was really only referenced at that point on the "new" cell accidentally
because the synapse ID it really meant to point to had changed.

2. Minor issue from last commit or so, the copied cell ID needs to be set
to the new synapse ID only after removing it from the network, when it
is newly created.