Fix three bugs by adding more thread synchronization during Train()
Fix 1: Each training thread's network was never getting updated from
the original during the course of training, so it could not benefit
from training that happened on other threads until the next entire
run. That is considered a bug. It is now fixed.

Fix 2: Updating the thread's network caused race condition crashes
due to cloning the original network back to the thread network
before the merge finished being applied.
A callback channel is being added to block the thread and keep
from continuing training before the merge was complete. Then the
original network can be cloned and overwrite the thread network.

Fix 3: A mutex was added to keep from cloning the original network
at the same time across threads. This is a separate issue because
we would end up having other threads trying to clone a network
at the same time, potentially during a merge of another thread,
but when the callback channel was not blocked.

It is worth noting that Fix 3 also fixed the issue of only being
able to train on (num cpus - 1) threads.