Tally cell voltage changes before firing cell
Add an extra operation where we tally all the firing synapse
effects before applying them to the cell they are firing.

Before this commit, it would be a race condition / random
if a cell fired due to an upstream synapse firing. A strong
synapse might immediately fire a cell, and block it. And all
other synapses afterward would not count until the next
Step() operation.

Now we will make a list of all the synapses that need to fire,
and sum all their effects. If that total sum is enough to
fire the cell, it will be fired. This should make the network
more predictable and take into account all synapses.