fix: nonexistent cells after master merge, concurrent firing pattern iteration+write
Vocabulary Inputs and Outputs were able to reference cell IDs that changed during a
network merge. This fixes the bug.

FiringPatterns are maps, and therefore pass-by-reference. But when cloning vocabulary,
we weren’t copying each value in the FiringPattern map, just the using map pointer
(well, maps are pointers). So Vocab.Inputs and Vocab.Outputs were shared across training
threads. It likely wasn’t a problem before because it was only recently made so Inputs
are modified on a thread, and FindClosestOutputCollection is called along with changing output FPs.