fix: nil synapses were causing havoc #40
deduping had been disabled, but is no longer.

network.Synapses sets synapses to nil after removing them, because now
instead of being able do delete a map entry, it is a slice[] which
cannot have an entry deleted. the nil synapse was breaking all kinds
of things, and just needed to be thoughtfully handled.

Also:
- less often deduping
- someday we should do #58 - reuse nil synapses